Switch to download tarballs from CloudFront
Instead of downloading tarballs from the S3 bucket, this switches to
download these from the CloudFront distribution. This is one missing
piece to enable building JAX in the prerelease workflows as the
prerelease buckets do not provide public read access and files are only
accessible through the CloudFront distribution by design.