[third-party] Add GMP to common sysdeps (#2560)
GMP is a required dependency for ROCgdb, so include it along with the
other sysdeps.

The sources are cached from the official website. We use the 6.3.0
release.

We do some light smoke testing to make sure the archive and libraries
get built.

Dependencies: GMP requires the texinfo package.

GMP is OS-agnostic, so is kept under third-party/sysdeps/common.

In the future it will be used for Windows-based builds of the
debug-tools.