[ci] Getting CI nightly back to green signal (#2621)
## Motivation

Currently, CI nightly and full tests are failing due to timeout errors.
Even though we provide generous timeouts, the tests continue to fail
frequently.

We are planning to get some signal from tests while providing a green
run, then investigating the issues in the background with issue #2616 .

Also removing test runner for `gfx1150 Linux` (Issue #2614) as it is
flaky and needs to investigate

## Technical Details

Updates to gtest filters for various components, in order to get test
times to a reasonable and non-flaky failures

## Test Plan

Testing via CI

## Test Result

- CI test run: https://github.com/ROCm/TheRock/actions/runs/20322565753
- CI in PR will also test

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.